Add files via upload
This Pune based startup, Urdhvam, is revolutionising the way the severe groundwater crisis can be solved in India. Rahul is an engineer and Vinit is a hydrogeology expert. They worked together to address the agrarian distress while taking into consideration the engineering, hydrology, and hydrogeology aspects to resolve the problem of drying borewells as they devised the BoreCharger.